Reference

Exported

Macros

DataToolkit

DataToolkit.@d_str — Macro

@d_str -> loaded data

Shorthand for loading a dataset in the default format, d"iris" is equivalent to read(dataset("iris")).

DataToolkit.@data_cmd — Macro

@data_cmd -> Data REPL command result

Proxy for running the command in the Data REPL, e.g. data`config set demo 1` is equivalent to data> config set demo 1.

DataToolkitBase

Missing docstring.

Missing docstring for DataToolkit.@import. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.@addpkg. Check Documenter's build log for details.

Functions

DataToolkit

Missing docstring.

Missing docstring for init. Check Documenter's build log for details.

Missing docstring.

Missing docstring for plugins. Check Documenter's build log for details.

DataToolkitBase

Missing docstring.

Missing docstring for dataset. Check Documenter's build log for details.

Missing docstring.

Missing docstring for loadcollection!. Check Documenter's build log for details.

Types

DataToolkitBase

Missing docstring.

Missing docstring for DataSet. Check Documenter's build log for details.

Unexported

Macros

DataToolkit

Missing docstring.

Missing docstring for DataToolkit.@addpkgs. Check Documenter's build log for details.

Functions

DataToolkitBase

Missing docstring.

Missing docstring for DataToolkit.getlayer. Check Documenter's build log for details.

Types

DataToolkitBase

Missing docstring.

Missing docstring for DataToolkit.DataCollection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataSet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataStorage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataLoader. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataWriter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.Identifier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.Plugin. Check Documenter's build log for details.