Reference

This is the public API for DataToolkit. Some symbols have been exported for convenience, others need to be specifically imported or accessed with DataToolkit.<thing>.

Exported Symbols

Macros

DataToolkit

DataToolkit.@d_str — Macro

@d_str -> loaded data

Shorthand for loading a dataset in the default format, d"iris" is equivalent to read(dataset("iris")).

DataToolkit.@data_cmd — Macro

@data_cmd -> Data REPL command result

Proxy for running the command in the Data REPL, e.g. data`config set demo 1` is equivalent to data> config set demo 1.

DataToolkitBase

Missing docstring.

Missing docstring for @import. Check Documenter's build log for details.

Functions

DataToolkitBase

Missing docstring.

Missing docstring for dataset. Check Documenter's build log for details.

Missing docstring.

Missing docstring for loadcollection!. Check Documenter's build log for details.

Types

DataToolkitBase

Missing docstring.

Missing docstring for DataSet. Check Documenter's build log for details.

Unexported Symbols

Modules

DataToolkitBase and DataToolkitCommon are available as Base and Common respectively.

Macros

DataToolkit

DataToolkit.@addpkgs — Macro

@addpkgs pkgs...

For each named package, register it with DataToolkitBase. Each package must be a dependency of the current module, recorded in its Project.toml.

This allows the packages to be used with DataToolkitBase.@import.

Instead of providing a list of packages, the symbol * can be provided to register all dependencies.

This must be run at runtime to take effect, so be sure to place it in the __init__ function of a package.

Examples

@addpkgs JSON3 CSV
@addpkgs * # Register all dependencies

Missing docstring.

Missing docstring for DataToolkit.@addpkg. Check Documenter's build log for details.

Functions

DataToolkit

DataToolkit.init — Function

init(mod::Module=Main; force::Bool=false)

Load the mod-local Data.toml if it exists. When mod is Main, every Data.toml on the load path is loaded. Unless force is set, the data collection is soft-loaded.

A Data.d directory can be used in place of a Data.toml, in which case every toml file within it will be read. Mixing Data.d/*.toml and Data.toml is discouraged.

DataToolkit.plugins — Function

plugins()

List the currently availible plugins, by name.

DataToolkit.addpkgs — Function

addpkgs(mod::Module, pkgs::Vector{Symbol})

For each package in pkgs, which are dependencies recorded in mod's Project.toml, register the package with DataToolkitBase.addpkg.

If pkgs consists of the single symbol :*, then all dependencies of mod will be registered.

This must be run at runtime to take effect, so be sure to place it in the __init__ function of a package.

DataToolkitBase

Missing docstring.

Missing docstring for DataToolkit.getlayer. Check Documenter's build log for details.

Types

DataToolkitBase

Missing docstring.

Missing docstring for DataToolkit.DataCollection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataSet. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataStorage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataLoader. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.DataWriter. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.Identifier. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DataToolkit.Plugin. Check Documenter's build log for details.